add new functions for getting object by vdm id
